+++ src/main/java/mcheli/ship/MCH_ShipPacketHandler.java
@@ -31,8 +31,11 @@ public class MCH_ShipPacketHandler {
                 mcheli.aircraft.MCH_BaseVehiclePacketHandler.handleVehicleAccessLockToggle(player, plane, pc);
             mcheli.aircraft.MCH_BaseVehiclePacketHandler.handleRadarToggle(player, plane, pc);
                 if(pc.isUnmount == 1) {
-                    plane.unmountEntity();
-                } else if(pc.isUnmount == 2) {
+                    if(mcheli.aircraft.MCH_BaseVehiclePacketHandler.validateNormalDismount(
+                          player, plane, pc.dismountMountEntityId, pc.dismountParentEntityId, pc.dismountSeatId)) {
+                       plane.unmountEntity();
+                    }
+                 } else if(pc.isUnmount == 2) {
                     plane.unmountCrew();
                 } else if(pc.ejectSeat) {
                     plane.ejectSeat(player);